LayoutTests/platform/chromium/http/tests/security/cross-frame-access-history-get-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-history-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-history-get.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-history-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-history-get.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-history-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-history-get.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-history-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-history-get.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-history-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-history-get.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-history-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-history-get.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-history-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-history-get.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe-for-history-get-test.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-history-get.html. Domains, protocols and ports must match.



----- tests for getting window.history and its properties -----

PASS: canGet('targetWindow.history.length') should be 'false' and is.
PASS: canGet('targetWindow.history.back') should be 'true' and is.
PASS: canGet('targetWindow.history.forward') should be 'true' and is.
PASS: canGet('targetWindow.history.go') should be 'true' and is.
*** FAIL: canGet('targetWindow.history.toString') should be 'true' but instead is false. ***
*** FAIL: toString('targetWindow.history') should be '[object History]' but instead is [exception]. ***
PASS: canGet('targetWindow.__proto__') should be 'false' and is.
PASS: canGet('targetWindow.constructor') should be 'false' and is.
PASS: canGet('targetWindow.history.existingCustomProperty') should be 'false' and is.
PASS: canGet('targetWindow.history.__proto__.prototypeCustomProperty') should be 'false' and is.
PASS: canCall('targetWindow.history.existingCustomFunction') should be 'false' and is.
PASS: canCall('targetWindow.history.prototypeCustomFunction') should be 'false' and is.
